refactor(Islands): report card part of build step
- make all methods purely functional
- invoke the script only once at the very end.
